pipeline/requirements.txt
google-api-python-client==2.118.0
python-dotenv==1.0.1
yt-dlp[default]
mlx-whisper
anthropic
